@@ -1,10 +1,10 @@
 name: signalr_netcore
 description: A flutter client for ASP.NET Core SignalR. The client is able to invoke server side hub functions and to receive method invocations issued by the server. 
-version: 1.1.1
+version: 1.2.0
 homepage: https://github.com/sefidgaran/signalr_client
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
 
