@@ -33,78 +33,77 @@ enum HubConnectionState {
 }
 
 typedef InvocationEventCallback = void Function(
-    HubMessageBase invocationEvent, Exception error);
-typedef MethodInvocationFunc = void Function(List<Object> arguments);
-typedef ClosedCallback = void Function({Exception error});
-typedef ReconnectingCallback = void Function({Exception error});
-typedef ReconnectedCallback = void Function({String connectionId});
+    HubMessageBase? invocationEvent, Exception? error);
+typedef MethodInvocationFunc = void Function(List<Object>? arguments);
+typedef ClosedCallback = void Function({Exception? error});
+typedef ReconnectingCallback = void Function({Exception? error});
+typedef ReconnectedCallback = void Function({String? connectionId});
 
 /// Represents a connection to a SignalR Hub
 class HubConnection {
   // Either a string (json) or Uint8List (binary);
-  Object _cachedPingMessage;
+  Object? _cachedPingMessage;
   final IConnection _connection;
-  final Logger _logger;
+  final Logger? _logger;
   final IRetryPolicy _reconnectPolicy;
   final IHubProtocol _protocol;
   final HandshakeProtocol _handshakeProtocol;
 
-  Map<String, InvocationEventCallback> _callbacks;
-  Map<String, List<MethodInvocationFunc>> _methods;
-  int _invocationId;
+  late Map<String?, InvocationEventCallback> _callbacks;
+  late Map<String, List<MethodInvocationFunc>> _methods;
+  int? _invocationId;
 
-  List<ClosedCallback> _closedCallbacks;
-  List<ReconnectingCallback> _reconnectingCallbacks;
-  List<ReconnectedCallback> _reconnectedCallbacks;
+  late List<ClosedCallback> _closedCallbacks;
+  late List<ReconnectingCallback> _reconnectingCallbacks;
+  late List<ReconnectedCallback> _reconnectedCallbacks;
 
-  bool _receivedHandshakeResponse;
-  Completer _handshakeCompleter;
-  Exception _stopDuringStartError;
+  late bool _receivedHandshakeResponse;
+  Completer? _handshakeCompleter;
+  Exception? _stopDuringStartError;
 
-  HubConnectionState _connectionState;
+  HubConnectionState? _connectionState;
 
   // connectionStarted is tracked independently from connectionState, so we can check if the
   // connection ever did successfully transition from connecting to connected before disconnecting.
-  bool _connectionStarted;
-  Future<void> _startPromise;
-  Future<void> _stopPromise;
+  late bool _connectionStarted;
+  Future<void>? _startPromise;
+  Future<void>? _stopPromise;
 
   // The type of these a) doesn't matter and b) varies when building in browser and node contexts
   // Since we're building the WebPack bundle directly from the TypeScript, this matters (previously
   // we built the bundle from the compiled JavaScript).
-  Timer _reconnectDelayTimer;
-  Timer _timeoutTimer;
-  Timer _pingServerTimer;
+  Timer? _reconnectDelayTimer;
+  Timer? _timeoutTimer;
+  Timer? _pingServerTimer;
 
   /// The server timeout in milliseconds.
   ///
   /// If this timeout elapses without receiving any messages from the server, the connection will be terminated with an error.
   /// The default timeout value is 30,000 milliseconds (30 seconds).
   ///
-  int serverTimeoutInMilliseconds;
+  late int serverTimeoutInMilliseconds;
 
   /// Default interval at which to ping the server.
   ///
   /// The default value is 15,000 milliseconds (15 seconds).
   /// Allows the server to detect hard disconnects (like when a client unplugs their computer).
   ///
-  int keepAliveIntervalInMilliseconds;
+  late int keepAliveIntervalInMilliseconds;
 
   /// Indicates the state of the {@link HubConnection} to the server.
-  HubConnectionState get state => _connectionState;
+  HubConnectionState? get state => _connectionState;
 
   /// Represents the connection id of the {@link HubConnection} on the server. The connection id will be null when the connection is either
   /// in the disconnected state or if the negotiation step was skipped.
-  String get connectionId =>
-      _connection != null ? _connection.connectionId : null;
+  String? get connectionId => _connection.connectionId ?? null;
 
   /// Indicates the url of the {@link HubConnection} to the server. */
-  String get baseUrl => _connection != null ? _connection.baseUrl : "";
+  String? get baseUrl => _connection.baseUrl ?? "";
 
   /// Sets a new url for the HubConnection. Note that the url can only be changed when the connection is in either the Disconnected or
   /// Reconnecting states.
   /// @param {string} url The url to connect to.
-  set baseUrl(String url) {
+  set baseUrl(String? url) {
     if (_connectionState != HubConnectionState.Disconnected &&
         _connectionState != HubConnectionState.Reconnecting) {
       throw GeneralError(
@@ -119,17 +118,15 @@ class HubConnection {
   }
 
   static HubConnection create(
-      IConnection connection, Logger logger, IHubProtocol protocol,
-      {IRetryPolicy reconnectPolicy}) {
+      IConnection connection, Logger? logger, IHubProtocol protocol,
+      {IRetryPolicy? reconnectPolicy}) {
     return HubConnection(connection, logger, protocol,
         reconnectPolicy: reconnectPolicy);
   }
 
-  HubConnection(IConnection connection, Logger logger, IHubProtocol protocol,
-      {IRetryPolicy reconnectPolicy})
-      : assert(connection != null),
-        assert(protocol != null),
-        _connection = connection,
+  HubConnection(IConnection connection, Logger? logger, IHubProtocol protocol,
+      {IRetryPolicy? reconnectPolicy})
+      : _connection = connection,
         _logger = logger,
         _protocol = protocol,
         _reconnectPolicy = reconnectPolicy ?? DefaultRetryPolicy(),
@@ -157,7 +154,7 @@ class HubConnection {
   ///
   /// Returns a Promise that resolves when the connection has been successfully established, or rejects with an error.
   ///
-  Future<void> start() async {
+  Future<void>? start() async {
     _startPromise = _startWithStateTransitions();
     return _startPromise;
   }
@@ -208,7 +205,7 @@ class HubConnection {
       _resetTimeoutPeriod();
       _resetKeepAliveInterval();
 
-      await _handshakeCompleter.future;
+      await _handshakeCompleter!.future;
 
       // It's important to check the stopDuringStartError instead of just relying on the handshakePromise
       // being rejected on close, because this continuation can run after both the handshake completed successfully
@@ -217,7 +214,7 @@ class HubConnection {
         // It's important to throw instead of returning a rejected promise, because we don't want to allow any state
         // transitions to occur between now and the calling code observing the exceptions. Returning a rejected promise
         // will cause the calling continuation to get scheduled to run later.
-        throw _stopDuringStartError;
+        throw _stopDuringStartError!;
       }
     } catch (e) {
       _logger?.finer(
@@ -228,7 +225,7 @@ class HubConnection {
 
       // HttpConnection.stop() should not complete until after the onclose callback is invoked.
       // This will transition the HubConnection to the disconnected state before HttpConnection.stop() completes.
-      await _connection.stop(error: e);
+      await _connection.stop(error: Exception(e));
       throw e;
     }
   }
@@ -252,7 +249,7 @@ class HubConnection {
     }
   }
 
-  Future<void> _stopInternal({Exception error}) async {
+  Future<void>? _stopInternal({Exception? error}) async {
     if (_connectionState == HubConnectionState.Disconnected) {
       _logger?.finer(
           "Call to HubConnection.stop($error) ignored because it is already in the disconnected state.");
@@ -300,13 +297,13 @@ class HubConnection {
   /// args: The arguments used to invoke the server method.
   /// Returns an object that yields results from the server as they are received.
   ///
-  Stream<Object> stream(String methodName, List<Object> args) {
+  Stream<Object?> stream(String methodName, List<Object> args) {
     final t = _replaceStreamingParams(args);
     final invocationDescriptor =
         _createStreamInvocation(methodName, args, t.item2);
 
-    Future<void> promiseQueue;
-    final StreamController streamController = StreamController<Object>(
+    late Future<void> promiseQueue;
+    final StreamController streamController = StreamController<Object?>(
       onCancel: () {
         final cancelInvocation =
             _createCancelInvocation(invocationDescriptor.invocationId);
@@ -317,7 +314,7 @@ class HubConnection {
     );
 
     _callbacks[invocationDescriptor.invocationId] =
-        (HubMessageBase invocationEvent, Exception error) {
+        (HubMessageBase? invocationEvent, Exception? error) {
       if (error != null) {
         streamController.addError(error);
         return;
@@ -345,7 +342,7 @@ class HubConnection {
     return streamController.stream;
   }
 
-  Future<void> _sendMessage(Object message) {
+  Future<void> _sendMessage(Object? message) {
     _resetKeepAliveInterval();
     return _connection.send(message);
   }
@@ -354,7 +351,7 @@ class HubConnection {
   /// message: The object to serialize and send.
   ///
   Future<void> _sendWithProtocol(Object message) {
-    return _sendMessage(_protocol.writeMessage(message));
+    return _sendMessage(_protocol.writeMessage(message as HubMessageBase));
   }
 
   /// Invokes a hub method on the server using the specified name and arguments. Does not wait for a response from the receiver.
@@ -366,7 +363,7 @@ class HubConnection {
   /// args: The arguments used to invoke the server method.
   /// Returns a Promise that resolves when the invocation has been successfully sent, or rejects with an error.
   ///
-  Future<void> send(String methodName, {List<Object> args}) {
+  Future<void> send(String methodName, {List<Object>? args}) {
     args = args ?? [];
     final t = _replaceStreamingParams(args);
     final sendPromise =
@@ -386,7 +383,7 @@ class HubConnection {
   /// args: The arguments used to invoke the server method.
   /// Returns a Future that resolves with the result of the server method (if any), or rejects with an error.
   ///
-  Future<Object> invoke(String methodName, {List<Object> args}) {
+  Future<Object> invoke(String methodName, {List<Object>? args}) {
     args = args ?? [];
     final t = _replaceStreamingParams(args);
     final invocationDescriptor =
@@ -395,7 +392,7 @@ class HubConnection {
     final completer = Completer<Object>();
 
     _callbacks[invocationDescriptor.invocationId] =
-        (HubMessageBase invocationEvent, Exception error) {
+        (HubMessageBase? invocationEvent, Exception? error) {
       if (error != null) {
         if (!completer.isCompleted) completer.completeError(error);
         return;
@@ -437,7 +434,7 @@ class HubConnection {
   /// newMethod: The handler that will be raised when the hub method is invoked.
   ///
   void on(String methodName, MethodInvocationFunc newMethod) {
-    if (isStringEmpty(methodName) || newMethod == null) {
+    if (isStringEmpty(methodName)) {
       return;
     }
 
@@ -447,11 +444,11 @@ class HubConnection {
     }
 
     // Preventing adding the same handler multiple times.
-    if (_methods[methodName].indexOf(newMethod) != -1) {
+    if (_methods[methodName]!.indexOf(newMethod) != -1) {
       return;
     }
 
-    _methods[methodName].add(newMethod);
+    _methods[methodName]!.add(newMethod);
   }
 
   /// Removes the specified handler for the specified hub method.
@@ -463,13 +460,13 @@ class HubConnection {
   /// method: The handler to remove. This must be the same Function instance as the one passed to {@link @microsoft/signalr.HubConnection.on}.
   /// If the method handler is omitted, all handlers for that method will be removed.
   ///
-  void off(String methodName, {MethodInvocationFunc method}) {
+  void off(String methodName, {MethodInvocationFunc? method}) {
     if (isStringEmpty(methodName)) {
       return;
     }
 
     methodName = methodName.toLowerCase();
-    final handlers = _methods[methodName];
+    final List<void Function(List<Object>)>? handlers = _methods[methodName];
     if (handlers == null) {
       return;
     }
@@ -492,9 +489,7 @@ class HubConnection {
   /// callback: The handler that will be invoked when the connection is closed. Optionally receives a single argument containing the error that caused the connection to close (if any).
   ///
   void onclose(ClosedCallback callback) {
-    if (callback != null) {
-      _closedCallbacks.add(callback);
-    }
+    _closedCallbacks.add(callback);
   }
 
   /// Registers a handler that will be invoked when the connection starts reconnecting.
@@ -502,9 +497,7 @@ class HubConnection {
   /// callback: The handler that will be invoked when the connection starts reconnecting. Optionally receives a single argument containing the error that caused the connection to start reconnecting (if any).
   ///
   onreconnecting(ReconnectingCallback callback) {
-    if (callback != null) {
-      _reconnectingCallbacks.add(callback);
-    }
+    _reconnectingCallbacks.add(callback);
   }
 
   /// Registers a handler that will be invoked when the connection successfully reconnects.
@@ -512,12 +505,10 @@ class HubConnection {
   /// callback: The handler that will be invoked when the connection successfully reconnects.
   ///
   onreconnected(ReconnectedCallback callback) {
-    if (callback != null) {
-      _reconnectedCallbacks.add(callback);
-    }
+    _reconnectedCallbacks.add(callback);
   }
 
-  void _processIncomingData(Object data) {
+  void _processIncomingData(Object? data) {
     _cleanupTimeout();
 
     if (!_receivedHandshakeResponse) {
@@ -534,12 +525,13 @@ class HubConnection {
         _logger?.finest("Handle message of type '${message.type}'.");
         switch (message.type) {
           case MessageType.Invocation:
-            _invokeClientMethod(message);
+            _invokeClientMethod(message as InvocationMessage);
             break;
           case MessageType.StreamItem:
           case MessageType.Completion:
             final invocationMsg = message as HubInvocationMessage;
-            final callback = _callbacks[invocationMsg.invocationId];
+            final void Function(HubMessageBase, Exception?)? callback =
+                _callbacks[invocationMsg.invocationId];
             if (callback != null) {
               if (message.type == MessageType.Completion) {
                 _callbacks.remove(invocationMsg.invocationId);
@@ -554,9 +546,9 @@ class HubConnection {
             _logger?.info("Close message received from server.");
             final closeMessage = message as CloseMessage;
 
-            final Exception error = closeMessage.error != null
+            final Exception? error = closeMessage.error != null
                 ? GeneralError(
-                    "Server returned an error on close: " + closeMessage.error)
+                    "Server returned an error on close: " + closeMessage.error!)
                 : null;
 
             if (closeMessage.allowReconnect == true) {
@@ -581,7 +573,7 @@ class HubConnection {
   }
 
   /// data is either a string (json) or a Uint8List (binary)
-  Object _processHandshakeResponse(Object data) {
+  Object? _processHandshakeResponse(Object? data) {
     ParseHandshakeResponseResult handshakeResult;
 
     try {
@@ -592,7 +584,7 @@ class HubConnection {
 
       final error = GeneralError(message);
 
-      if (!_handshakeCompleter.isCompleted) {
+      if (!_handshakeCompleter!.isCompleted) {
         _handshakeCompleter?.completeError(error);
       }
       _handshakeCompleter = null;
@@ -605,7 +597,7 @@ class HubConnection {
 
       final error = GeneralError(message);
 
-      if (!_handshakeCompleter.isCompleted) {
+      if (!_handshakeCompleter!.isCompleted) {
         _handshakeCompleter?.completeError(error);
       }
       _handshakeCompleter = null;
@@ -614,7 +606,7 @@ class HubConnection {
       _logger?.finer("Server handshake complete.");
     }
 
-    if (!_handshakeCompleter.isCompleted) _handshakeCompleter?.complete();
+    if (!_handshakeCompleter!.isCompleted) _handshakeCompleter?.complete();
     _handshakeCompleter = null;
     return handshakeResult.remainingData;
   }
@@ -639,8 +631,8 @@ class HubConnection {
   void _resetTimeoutPeriod() {
     _cleanupTimeout();
     if ((_connection.features == null) ||
-        (_connection.features.inherentKeepAlive == null) ||
-        (!_connection.features.inherentKeepAlive)) {
+        (_connection.features!.inherentKeepAlive == null) ||
+        (!_connection.features!.inherentKeepAlive!)) {
       // Set the timeout timer
       _timeoutTimer = Timer.periodic(
           Duration(milliseconds: serverTimeoutInMilliseconds), _serverTimeout);
@@ -656,7 +648,7 @@ class HubConnection {
   }
 
   void _invokeClientMethod(InvocationMessage invocationMessage) {
-    final methods = _methods[invocationMessage.target.toLowerCase()];
+    final methods = _methods[invocationMessage.target!.toLowerCase()];
     if (methods != null) {
       methods.forEach((m) => m(invocationMessage.arguments));
       if (!isStringEmpty(invocationMessage.invocationId)) {
@@ -674,7 +666,7 @@ class HubConnection {
     }
   }
 
-  void _connectionClosed({Exception error}) {
+  void _connectionClosed({Exception? error}) {
     _logger?.finer(
         "HubConnection.connectionClosed($error) called while in state $_connectionState.");
 
@@ -687,7 +679,7 @@ class HubConnection {
     // If the handshake is in progress, start will be waiting for the handshake promise, so we complete it.
     // If it has already completed, this should just noop.
     if (_handshakeCompleter != null) {
-      if (!_handshakeCompleter.isCompleted) _handshakeCompleter.complete();
+      if (!_handshakeCompleter!.isCompleted) _handshakeCompleter!.complete();
     }
 
     _cancelCallbacksWithError(error ??
@@ -699,8 +691,7 @@ class HubConnection {
 
     if (_connectionState == HubConnectionState.Disconnecting) {
       _completeClose(error: error);
-    } else if (_connectionState == HubConnectionState.Connected &&
-        _reconnectPolicy != null) {
+    } else if (_connectionState == HubConnectionState.Connected) {
       _reconnect(error: error);
     } else if (_connectionState == HubConnectionState.Connected) {
       _completeClose(error: error);
@@ -713,7 +704,7 @@ class HubConnection {
     // 3. The Disconnected state in which case we're already done.
   }
 
-  _completeClose({Exception error}) {
+  _completeClose({Exception? error}) {
     if (_connectionStarted) {
       _connectionState = HubConnectionState.Disconnected;
       _connectionStarted = false;
@@ -727,7 +718,7 @@ class HubConnection {
     }
   }
 
-  _reconnect({Exception error}) async {
+  _reconnect({Exception? error}) async {
     final reconnectStartTime = DateTime.now();
     var previousReconnectAttempts = 0;
     Exception retryError = error != null
@@ -802,7 +793,7 @@ class HubConnection {
           return;
         }
 
-        retryError = e;
+        retryError = Exception(e);
         nextRetryDelay = _getNextRetryDelay(
             previousReconnectAttempts++,
             DateTime.now().difference(reconnectStartTime).inMilliseconds,
@@ -816,7 +807,7 @@ class HubConnection {
     _completeClose();
   }
 
-  int _getNextRetryDelay(
+  int? _getNextRetryDelay(
       int previousRetryCount, int elapsedMilliseconds, Exception retryReason) {
     try {
       return _reconnectPolicy.nextRetryDelayInMilliseconds(
@@ -829,7 +820,8 @@ class HubConnection {
   }
 
   _cancelCallbacksWithError(Exception error) {
-    final callbacks = _callbacks;
+    final Map<String?, void Function(HubMessageBase?, Exception)> callbacks =
+        _callbacks;
     _callbacks = {};
 
     callbacks.forEach((_, value) => {value(null, error)});
@@ -837,21 +829,21 @@ class HubConnection {
 
   void _cleanupPingTimer() {
     if (_pingServerTimer != null) {
-      _pingServerTimer.cancel();
+      _pingServerTimer!.cancel();
       _pingServerTimer = null;
     }
   }
 
   void _cleanupTimeout() {
     if (_timeoutTimer != null) {
-      _timeoutTimer.cancel();
+      _timeoutTimer!.cancel();
       _timeoutTimer = null;
     }
   }
 
   void _cleanupReconnectTimer() {
     if (_reconnectDelayTimer != null) {
-      _reconnectDelayTimer.cancel();
+      _reconnectDelayTimer!.cancel();
       _reconnectDelayTimer = null;
     }
   }
@@ -863,14 +855,14 @@ class HubConnection {
           methodName, args, streamIds, MessageHeaders(), null);
     } else {
       final invocationId = _invocationId;
-      _invocationId++;
+      _invocationId = _invocationId! + 1;
 
       return InvocationMessage(methodName, args, streamIds, MessageHeaders(),
           invocationId.toString());
     }
   }
 
-  _launchStreams(List<Stream<Object>> streams, Future<void> promiseQueue) {
+  _launchStreams(List<Stream<Object>> streams, Future<void>? promiseQueue) {
     if (streams.length == 0) {
       return;
     }
@@ -883,10 +875,10 @@ class HubConnection {
     // We want to iterate over the keys, since the keys are the stream ids
     for (var i = 0; i < streams.length; i++) {
       streams[i].listen((item) {
-        promiseQueue = promiseQueue.then((_) =>
+        promiseQueue = promiseQueue?.then((_) =>
             _sendWithProtocol(_createStreamItemMessage(i.toString(), item)));
       }, onDone: () {
-        promiseQueue = promiseQueue.then(
+        promiseQueue = promiseQueue?.then(
             (_) => _sendWithProtocol(_createCompletionMessage(i.toString())));
       }, onError: (err) {
         String message;
@@ -896,7 +888,7 @@ class HubConnection {
           message = "Unknown error";
         }
 
-        promiseQueue = promiseQueue.then((_) => _sendWithProtocol(
+        promiseQueue = promiseQueue?.then((_) => _sendWithProtocol(
             _createCompletionMessage(i.toString(), error: message)));
       });
     }
@@ -910,10 +902,10 @@ class HubConnection {
     for (var i = 0; i < args.length; i++) {
       final argument = args[i];
       if (argument is Stream) {
-        final streamId = _invocationId;
-        _invocationId++;
+        final streamId = _invocationId!;
+        _invocationId = _invocationId! + 1;
         // Store the stream for later use
-        streams[streamId] = argument;
+        streams[streamId] = argument as Stream<Object>;
         streamIds.add(streamId.toString());
 
         // remove stream from args
@@ -929,13 +921,13 @@ class HubConnection {
   StreamInvocationMessage _createStreamInvocation(
       String methodName, List<Object> args, List<String> streamIds) {
     final invocationId = _invocationId;
-    _invocationId++;
+    _invocationId = _invocationId! + 1;
 
     return StreamInvocationMessage(
         methodName, args, streamIds, MessageHeaders(), invocationId.toString());
   }
 
-  CancelInvocationMessage _createCancelInvocation(String id) {
+  CancelInvocationMessage _createCancelInvocation(String? id) {
     return CancelInvocationMessage(new MessageHeaders(), id);
   }
 
@@ -944,9 +936,10 @@ class HubConnection {
   }
 
   CompletionMessage _createCompletionMessage(String id,
-      {Object error, Object result}) {
+      {Object? error, Object? result}) {
     if (error != null) {
-      return CompletionMessage(error, null, new MessageHeaders(), id);
+      return CompletionMessage(
+          error as String?, null, new MessageHeaders(), id);
     }
 
     return CompletionMessage(null, result, new MessageHeaders(), id);
