@@ -15,17 +15,17 @@ import 'default_reconnect_policy.dart';
 class HubConnectionBuilder {
   // Properties
 
-  IHubProtocol _protocol;
+  IHubProtocol? _protocol;
 
-  HttpConnectionOptions _httpConnectionOptions;
+  HttpConnectionOptions? _httpConnectionOptions;
 
-  String _url;
+  String? _url;
 
-  Logger _logger;
+  Logger? _logger;
 
   /// If defined, this indicates the client should automatically attempt to reconnect if the connection is lost. */
   ///
-  IRetryPolicy _reconnectPolicy;
+  IRetryPolicy? _reconnectPolicy;
 
   /// Configures console logging for the HubConnection.
   ///
@@ -46,7 +46,7 @@ class HubConnectionBuilder {
   /// Returns the builder instance, for chaining.
   ///
   HubConnectionBuilder withUrl(String url,
-      {HttpConnectionOptions options, HttpTransportType transportType}) {
+      {HttpConnectionOptions? options, HttpTransportType? transportType}) {
     assert(!isStringEmpty(url));
     assert(!(options != null && transportType != null));
 
@@ -66,14 +66,13 @@ class HubConnectionBuilder {
   /// protocol: The IHubProtocol implementation to use.
   ///
   HubConnectionBuilder withHubProtocol(IHubProtocol protocol) {
-    assert(protocol != null);
 
     _protocol = protocol;
     return this;
   }
 
   HubConnectionBuilder withAutomaticReconnect(
-      {IRetryPolicy reconnectPolicy, List<int> retryDelays}) {
+      {IRetryPolicy? reconnectPolicy, List<int>? retryDelays}) {
     assert(_reconnectPolicy == null);
 
     if (reconnectPolicy == null && retryDelays == null) {
@@ -102,7 +101,7 @@ class HubConnectionBuilder {
       throw new GeneralError(
           "The 'HubConnectionBuilder.withUrl' method must be called before building the connection.");
     }
-    final connection = HttpConnection(_url, options: httpConnectionOptions);
+    final connection = HttpConnection(_url!, options: httpConnectionOptions);
     return HubConnection.create(
         connection, _logger, _protocol ?? JsonHubProtocol(),
         reconnectPolicy: _reconnectPolicy);
