@@ -4,24 +4,24 @@ import "itransport.dart";
 
 class ConnectionFeatures {
   // Properties
-  bool inherentKeepAlive;
+  bool? inherentKeepAlive;
 
   // Methods
   ConnectionFeatures(this.inherentKeepAlive);
 }
 
 abstract class IConnection {
-  ConnectionFeatures features;
-  String connectionId;
+  ConnectionFeatures? features;
+  String? connectionId;
 
-  String baseUrl;
+  String? baseUrl;
 
-  Future<void> start({TransferFormat transferFormat});
-  Future<void> send(Object data);
-  Future<void> stop({Exception error});
+  Future<void> start({TransferFormat? transferFormat});
+  Future<void> send(Object? data);
+  Future<void>? stop({Exception? error});
 
-  OnReceive onreceive;
-  OnClose onclose;
+  OnReceive? onreceive;
+  OnClose? onclose;
 
   IConnection() : features = ConnectionFeatures(null);
 }
