@@ -11,52 +11,52 @@ import 'utils.dart';
 class WebSocketTransport implements ITransport {
   // Properties
 
-  Logger _logger;
-  AccessTokenFactory _accessTokenFactory;
+  Logger? _logger;
+  AccessTokenFactory? _accessTokenFactory;
   bool _logMessageContent;
-  WebSocketChannel _webSocket;
-  StreamSubscription<Object> _webSocketListenSub;
+  WebSocketChannel? _webSocket;
+  StreamSubscription<Object?>? _webSocketListenSub;
 
   @override
-  OnClose onClose;
+  OnClose? onClose;
 
   @override
-  OnReceive onReceive;
+  OnReceive? onReceive;
 
   // Methods
-  WebSocketTransport(AccessTokenFactory accessTokenFactory, Logger logger,
+  WebSocketTransport(AccessTokenFactory? accessTokenFactory, Logger? logger,
       bool logMessageContent)
       : this._accessTokenFactory = accessTokenFactory,
         this._logger = logger,
         this._logMessageContent = logMessageContent;
 
   @override
-  Future<void> connect(String url, TransferFormat transferFormat) async {
+  Future<void> connect(String? url, TransferFormat transferFormat) async {
     assert(url != null);
-    assert(transferFormat != null);
 
     _logger?.finest("(WebSockets transport) Connecting");
 
     if (_accessTokenFactory != null) {
-      final token = await _accessTokenFactory();
+      final token = await _accessTokenFactory!();
       if (!isStringEmpty(token)) {
         final encodedToken = Uri.encodeComponent(token);
-        url +=
-            (url.indexOf("?") < 0 ? "?" : "&") + "access_token=$encodedToken";
+        url = url! +
+            (url.indexOf("?") < 0 ? "?" : "&") +
+            "access_token=$encodedToken";
       }
     }
 
     var websocketCompleter = Completer();
     var opened = false;
-    url = url.replaceFirst('http', 'ws');
+    url = url!.replaceFirst('http', 'ws');
     _logger?.finest("WebSocket try connecting to '$url'.");
     _webSocket = WebSocketChannel.connect(Uri.parse(url));
     opened = true;
     if (!websocketCompleter.isCompleted) websocketCompleter.complete();
     _logger?.info("WebSocket connected to '$url'.");
-    _webSocketListenSub = _webSocket.stream.listen(
+    _webSocketListenSub = _webSocket!.stream.listen(
       // onData
-      (Object message) {
+      (Object? message) {
         if (_logMessageContent && message is String) {
           _logger?.finest(
               "(WebSockets transport) data received. message ${getDataDetail(message, _logMessageContent)}.");
@@ -65,7 +65,7 @@ class WebSocketTransport implements ITransport {
         }
         if (onReceive != null) {
           try {
-            onReceive(message);
+            onReceive!(message);
           } catch (error) {
             _logger?.severe(
                 "(WebSockets transport) error calling onReceive, error: $error");
@@ -75,7 +75,7 @@ class WebSocketTransport implements ITransport {
       },
 
       // onError
-      onError: (Object error) {
+      onError: (Object? error) {
         var e = error != null ? error : "Unknown websocket error";
         if (!websocketCompleter.isCompleted) {
           websocketCompleter.completeError(e);
@@ -88,7 +88,7 @@ class WebSocketTransport implements ITransport {
         // We'll reject the connect call instead
         if (opened) {
           if (onClose != null) {
-            onClose();
+            onClose!();
           }
         } else {
           if (!websocketCompleter.isCompleted) {
@@ -110,9 +110,9 @@ class WebSocketTransport implements ITransport {
       //_logger?.finest("(WebSockets transport) sending data.");
 
       if (data is String) {
-        _webSocket.sink.add(data);
+        _webSocket!.sink.add(data);
       } else if (data is Uint8List) {
-        _webSocket.sink.add(data);
+        _webSocket!.sink.add(data);
       } else {
         throw GeneralError("Content type is not handled.");
       }
@@ -132,17 +132,16 @@ class WebSocketTransport implements ITransport {
   _close() async {
     if (_webSocket != null) {
       // Clear websocket handlers because we are considering the socket closed now
-      if (_webSocketListenSub != null) {
-        await _webSocketListenSub.cancel();
-        _webSocketListenSub = null;
-      }
-      _webSocket.sink.close();
+      await _webSocketListenSub?.cancel();
+      _webSocketListenSub = null;
+
+      _webSocket!.sink.close();
       _webSocket = null;
     }
 
     _logger?.finest("(WebSockets transport) socket closed.");
     if (onClose != null) {
-      onClose();
+      onClose!();
     }
   }
 }
